fix(ferries): add schedules/routes endpoints
