+++ example/vessels/main.go
@@ -51,4 +51,20 @@ func main() {
 		fmt.Println(vesselLocations[1].VesselName)
 		fmt.Printf("%f°N, %f°W\n", vesselLocations[1].Latitude, vesselLocations[1].Longitude)
 	}
+
+	// get the route schedule
+	schedules, err := ferriesClient.GetRouteSchedules()
+	if err != nil {
+		panic(err)
+	}
+
+	for _, schedule := range schedules {
+		fmt.Printf("Schedule ID: %d, Route ID: %d, Description: %s\n", schedule.ScheduleID, schedule.RouteID, schedule.Description)
+		allSailings, err := ferriesClient.GetSchedulesTodayByRouteID(schedule.RouteID, true)
+		if err != nil {
+			panic(err)
+		}
+		fmt.Printf("last sailing: %s\n", allSailings.TerminalCombos[0].Times[len(allSailings.TerminalCombos[0].Times)-1].DepartingTime)
+	}
+
 }
